Extract the method for uploading an image
