+++ src/flickypedia/apis/__init__.py
@@ -1,6 +1,14 @@
-from .wikimedia import MissingFileException, UnknownWikimediaApiException, WikimediaApi
+from .wikimedia import (
+    DuplicateFilenameUploadException,
+    DuplicatePhotoUploadException,
+    MissingFileException,
+    UnknownWikimediaApiException,
+    WikimediaApi,
+)
 
 __all__ = [
+    "DuplicateFilenameUploadException",
+    "DuplicatePhotoUploadException",
     "MissingFileException",
     "UnknownWikimediaApiException",
     "WikimediaApi",
